% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fsl_indicators.R
\name{calculate_fsl_indicators}
\alias{calculate_fsl_indicators}
\title{Calculate FSL Indicators}
\usage{
calculate_fsl_indicators(
  data,
  FCSStap = NULL,
  FCSPulse = NULL,
  FCSPr = NULL,
  FCSVeg = NULL,
  FCSFruit = NULL,
  FCSDairy = NULL,
  FCSFat = NULL,
  FCSSugar = NULL,
  cutoff = "Cat28",
  rCSILessQlty = NULL,
  rCSIBorrow = NULL,
  rCSIMealSize = NULL,
  rCSIMealAdult = NULL,
  rCSIMealNb = NULL,
  HHhSNoFood_FR = NULL,
  HHhSBedHung_FR = NULL,
  HHhSNotEat_FR = NULL,
  HDDSStapCer = NULL,
  HDDSStapRoot = NULL,
  HDDSVeg = NULL,
  HDDSFruit = NULL,
  HDDSPrMeat = NULL,
  HDDSPrEgg = NULL,
  HDDSPrFish = NULL,
  HDDSPulse = NULL,
  HDDSDairy = NULL,
  HDDSFat = NULL,
  HDDSSugar = NULL,
  HDDSCond = NULL
)
}
\arguments{
\item{HDDSCond}{}
}
\description{
Calculate FSL Indicators
}
