% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{check_file_layout}
\alias{check_file_layout}
\title{Check a File's Layout}
\usage{
check_file_layout(path,
  max_file_length = get_cleanr_options("max_file_length"),
  max_file_width = get_cleanr_options("max_file_width"))
}
\arguments{
\item{path}{A path to a file, e.g. "checks.R".}

\item{max_file_length}{See \code{\link{check_file_length}}.}

\item{max_file_width}{See \code{\link{check_file_width}}.}
}
\value{
invisible(TRUE), but see \emph{Details}.
}
\description{
Check for number of lines and width of lines.
}
\details{
In case of a fail the function \code{\link{throw}}s a
condition of class c("cleanr", "error", "condition").
}
\examples{
print(check_file_layout(system.file("source", "R", "checks.R",
                                    package = "cleanr")))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

