% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{check_function_layout}
\alias{check_function_layout}
\title{check a function's layout}
\usage{
check_function_layout(object,
  max_lines_of_code = get_cleanr_options("max_lines_of_code"),
  max_lines = get_cleanr_options("max_lines"),
  max_arguments = get_cleanr_options("max_arguments"),
  max_nesting_depth = get_cleanr_options("max_nesting_depth"),
  max_line_width = get_cleanr_options("max_line_width"),
  check_return = get_cleanr_options("check_return"))
}
\arguments{
\item{object}{The function to be checked.}

\item{max_lines_of_code}{the maximum number of lines of code accepted.}

\item{max_lines}{the maximum number of lines accepted.}

\item{max_arguments}{the maximum number of arguments accepted.}

\item{max_nesting_depth}{the maximum nesting depth accepted.}

\item{max_line_width}{the maximum line width accepted.}

\item{check_return}{Should the function be tested to have an explicit return?}
}
\value{
invisible(TRUE), but see \emph{Details}.
}
\description{
run all \code{\link{function_checks}} on a function.
}
\details{
The functions catches the messages of "cleanr"-conditions
\code{\link{throw}}n by \code{\link{function_checks}} and, if it caught any,
\code{\link{throw}}s them.
}
\examples{
print(check_function_layout(check_num_lines))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

