% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrappers.r
\name{check_functions_in_file}
\alias{check_functions_in_file}
\title{check all functions defined in a file}
\usage{
check_functions_in_file(path, ...)
}
\arguments{
\item{path}{A path to a file, e.g. "checks.r".}

\item{...}{Argments to be passed to \code{\link{check_function_layout}}.}
}
\value{
invisible(TRUE), but see \emph{Details}.
}
\description{
run all \code{\link{check_function_layout}} on a file.
}
\details{
The functions catches the messages of 'coldr'-conditions \code{\link{throw}}n
by \code{\link{check_function_layout}} and, if it caught any,
\code{\link{throw}}s
them.
}
\section{Version}{
 $Id: 01015ff091d53e47fc1caa95805585b6e3911ba5 $
}
\examples{
print(check_functions_in_file(system.file('source', 'R', 'utils.r',
                                    package = 'coldr')))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

