% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Gen_sp.R
\name{check_Gen_sp}
\alias{check_Gen_sp}
\title{check_sp}
\usage{
check_Gen_sp(species_tesaurus, Gen_sp, k = 2)
}
\arguments{
\item{k:}{number of changes allowed by the fuzzy matching function.}
}
\description{
check_sp
}
