% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_responses.R
\name{check_other_responses}
\alias{check_other_responses}
\title{Check other responses}
\usage{
check_other_responses(data, other_columns)
}
\arguments{
\item{data}{this should be survey data in XML value format}

\item{other_columns}{list of all other responses variables, usually prepare this list from the XLSX tool}
}
\description{
This function will flag all other responses provided in the data, make sure to have a list of all other column.
}
