% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sp.R
\name{check_sp}
\alias{check_sp}
\title{check_sp}
\usage{
check_sp(template, Genus, Species, k = 2)
}
\arguments{
\item{k:}{number of changes allowed by the fuzzy matching function.}
}
\value{

}
\description{
check_sp
}
