% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_species.R
\name{clean_species}
\alias{clean_species}
\title{Helper function to clean taxonomy.}
\usage{
clean_species(species, verbose = TRUE, db = "itis")
}
\arguments{
\item{species}{A vector of species to check}
}
\value{
data.frame of original and corrected species.
}
\description{
This correct misspellings, synonyms and flags not recognizes species.
}
\examples{
\dontrun{
species <- c("Osmia rufa", "Osmia bicornis", "Osmia ruffa",
           "Osmia wikifluqie", "watermelon pie", "Osmia sp.")
clean_species(species)
}
}
