% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning_folders.R
\name{cleaning_folders}
\alias{cleaning_folders}
\title{cleaning folders}
\usage{
cleaning_folders()
}
\description{
This function will create a working directory for you, first it will check if you have a right working directory, and if so, will return message saying good to go. otherwise, it will crate structured working directory for you.
}
