% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_to_numeric.R
\name{coerce_numeric_columns}
\alias{coerce_numeric_columns}
\title{Coerce Numeric Columns}
\usage{
coerce_numeric_columns(data, numeric_columns)
}
\arguments{
\item{data}{A data frame containing the dataset.}

\item{numeric_columns}{A character vector of column names expected to be numeric.}
}
\value{
A data frame with specified columns coerced to numeric, changes made, and problematic rows dropped.
}
\description{
This function takes a column that a user assumes should have numeric inputs and
identifies any values that have been stored as text.
This function coerces them to numeric, drops any rows that have been coerced to "NA", and
prints the changes in a dataframe.
}
\examples{
updated_data <- coerce_numeric_columns(plants_clean, numeric_columns = c("X10_9_weight"))
}
