% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{coldr-package}
\alias{coldr-package}
\title{coding less dreadfully}
\description{
Check your R code for some of the most common layout flaws.
}
\details{
Many tried to teach us how to write code less dreadful, be it implicitly as
\cite{B. W. Kernighan and D. M. Ritchie. The C Programming Language, Second
Edition. Prentice-Hall, Englewood Cliffs, New Jersey, 1988}, be it
explicitly as
\cite{R.C. Martin. Clean Code: A Handbook of Agile Software Craftsmanship.
Pearson Education, 2008} or \cite{K. Passig and J. Jander. Weniger schlecht
programmieren. O'Reilly, 2013}.
So you should check your code for files too long or wide, functions with too
many lines, too wide lines, too many arguments or nestings too deep.
}
\section{Version}{
 $Id: 21e3a57157dcd729e5bbed73ed5d84159e8ea554 $
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
\seealso{
packages
\href{http://cran.r-project.org/web/packages/codetools/index.html}{\pkg{codetools}},
\href{http://cran.r-project.org/web/packages/lintr/index.html}{\pkg{lintr}}
and
\href{http://cran.r-project.org/web/packages/formatR/index.html}{\pkg{formatR}}.
}
\keyword{package}

