% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fsl_indicators.R
\name{compute_fcs}
\alias{compute_fcs}
\title{Compute FCS Indicator}
\usage{
compute_fcs(
  data,
  FCSStap = "FCSStap",
  FCSPulse = "FCSPulse",
  FCSPr = "FCSPr",
  FCSVeg = "FCSVeg",
  FCSFruit = "FCSFruit",
  FCSDairy = "FCSDairy",
  FCSFat = "FCSFat",
  FCSSugar = "FCSSugar",
  cutoff = "Cat28"
)
}
\arguments{
\item{data}{Survey data with FCS indicator as per the WFP Survey Designer}

\item{FCSStap}{Cereals, grains, roots and tubers}

\item{FCSPulse}{Pulses/ legumes / nuts}

\item{FCSPr}{Milk and other dairy products}

\item{FCSVeg}{Meat, fish and eggs}

\item{FCSFruit}{Vegetables and leaves}

\item{FCSDairy}{Fruits}

\item{FCSFat}{Oil/fat/butter}

\item{FCSSugar}{Sugar, or sweet}

\item{cutoff}{threshold}
}
\description{
This function will calculate the FCS indicator, you'll need to include the cutoff threshold in the function.
}
