% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{Survey Data}
\format{
A data frame with 100 rows and 17 variables:
\describe{
  \item{start}{survey start time}
  \item{end}{survey end time}
  \item{today}{date of data collection}
  \item{consent}{survey consent}
  \item{enum_name}{enumerator name}
  \item{region}{admin 1 level name}
  \item{district}{admin 2 level name}
  \item{ki_age}{respondent age}
  \item{average_monthly_income}{average household monhtly income}
  \item{belong_idp}{respondent belong to idp}
  \item{camp_structure}{is there camp structure}
  \item{camp_structure.no}{there is no camp structure}
  \item{camp_structure.yes_camp}{yes there camp committee}
  \item{camp_structure.yes_committee}{yes committe exist}
  \item{camp_structure.other}{other camp structures}
  \item{camp_structure_other}{other specefication}
  \item{uuid}{uuid}

}
}
\source{
Dumpy data generated
}
\usage{
data
}
\description{
A dataset containing household level data
}
\examples{
data(data)
}
\keyword{datasets}
