% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_template.R
\name{define_template}
\alias{define_template}
\title{define_template}
\usage{
define_template(example, species_tesaurus)
}
\arguments{
\item{species_tesaurus}{a speices checklist (data frame with columns Genus and Species)}

\item{example:}{a model data.frame with all possible categories and min and max numeric values allowed.}
}
\value{
a list
}
\description{
define_template
}
