% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invalid_lines.R
\name{detect_invalid_lines}
\alias{detect_invalid_lines}
\title{Detect Invalid 'line' Values in a Dataset}
\usage{
detect_invalid_lines(data)
}
\arguments{
\item{data}{A data frame that contains the dataset with a \code{line} column.}
}
\value{
A data frame with rows where the \code{line} column contains invalid values, along with suggested corrections.
}
\description{
This function is specific to a feature of my research and my plants_clean dataset. It identifies rows in the dataset where the values in the \code{line} column
deviate from the expected pattern (i.e., "C1", "SC1", or "SI1"). It returns a
data frame with the rows that contain invalid \code{line} values and suggests possible corrections.
}
\examples{
# Example dataset
plantdata <- data.frame(
  treatment = c("d", "d", "d", "wet", "d", "w"),
  pot_id = c(1, 2, 3, 4, 5, 6),
  line = c("C1", "SXX1", "SC1", "SI1", "C1", "C1"),
  mom = c(86, 225, 231, 672, 2, 27),
  dad = c(59, 258, 245, 697, 44, 100),
  tray_id = c(1, 10, 1, 1, 1, 1),
  dry_weight = c(165, 168, 158, 170, 172, 160)
)

# Detect invalid 'line' values
invalid_lines <- detect_invalid_lines(plantdata)
print(invalid_lines)

}
