% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanR.R
\name{dupliDrop}
\alias{dupliDrop}
\title{Remove columns with the same names}
\usage{
dupliDrop(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame without columns with the same names.
}
\description{
Remove columns with the same names
}
\examples{
dupliDrop(mtcars)
}
