% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pieces.R
\name{extract_pieces}
\alias{extract_pieces}
\title{extract_pieces}
\usage{
extract_pieces(to_split, pattern_ = "_", i = 1, f = 0, k = 1,
  subgenus = FALSE, species = FALSE, single = TRUE)
}
\arguments{
\item{single}{}
}
\description{
extract_pieces
}
