% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{file_checks}
\alias{check_file_length}
\alias{check_file_width}
\alias{file_checks}
\title{file checks}
\usage{
check_file_width(path, maximum = get_cleanr_options("max_file_width"))

check_file_length(path, maximum = get_cleanr_options("max_line_width"))
}
\arguments{
\item{path}{The path to the file to be checked.}

\item{maximum}{The maximum against which the file is to be tested.}
}
\value{
invisible(TRUE), but see \emph{Details}.
}
\description{
A set of tiny functions to check that files adhere to a layout style.
}
\details{
A file should have a clear layout, it should
\itemize{
  \item mot have too many lines and
  \item not have lines too wide.
}
At least this is what I think. Well, some others too.

All of the functions test whether their requirement is met. 
In case of a fail all \code{\link{throw}} a
condition of class c("cleanr", "error", "condition").
}
\examples{
print(check_file_width(system.file("source", "R", "checks.R",
                                    package = "cleanr")))
print(check_file_length(system.file("source", "R", "checks.R",
                                    package = "cleanr"), maximum = 300))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

