% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.r
\name{function_checks}
\alias{check_line_width}
\alias{check_nesting_depth}
\alias{check_num_arguments}
\alias{check_num_lines}
\alias{check_num_lines_of_code}
\alias{check_return}
\alias{function_checks}
\title{function checks}
\usage{
check_num_arguments(object, maximum = get_cleanr_options("max_arguments"))

check_nesting_depth(object, maximum = get_cleanr_options("max_nesting_depth"))

check_num_lines(object, maximum = get_cleanr_options("max_lines"))

check_num_lines_of_code(object,
  maximum = get_cleanr_options("max_lines_of_code"))

check_line_width(object, maximum = get_cleanr_options("max_line_width"))

check_return(object)
}
\arguments{
\item{object}{The function to be checked.
Should have been sourced with keep.source = TRUE (see
\code{\link{get_function_body}}.}

\item{maximum}{The maximum against which the function is to be tested.}
}
\value{
invisible(TRUE), but see \emph{Details}.
}
\description{
A set of tiny functions to check that functions adhere to a layout style.
}
\details{
A function should have a clear layout, it should
\itemize{
  \item not have too many arguments,
  \item not have nestings too deep,
  \item neither have too many lines nor
  \item have too many lines of code,
  \item not have lines too wide and
  \item explicitly \code{\link{return}} an object.
}
At least this is what I think. Well, some others too.

All of the functions test whether their requirement is met (some layout
feature such as number of arguments, nesting depth, line width is not greater
than the maximum given). In case of a fail all \code{\link{throw}} a
condition of class c("cleanr", "error", "condition").
}
\section{Warning}{
 \code{\link{check_return}} just \code{\link{grep}}s for a
for a line starting with a \code{\link{return}} statemtent (ah, see the code
for the real thing).
This doesn't ensure that \emph{all} \code{\link{return}} paths from the
function are explicit and it may miss a \code{\link{return}} path after a
semicolon.
It just checks if you use \code{\link{return}} at all.
}

\section{Version}{
 $Id: 01015ff091d53e47fc1caa95805585b6e3911ba5 $
}
\examples{
print(check_num_arguments(check_num_arguments))
print(check_nesting_depth(check_nesting_depth))
print(check_num_lines(check_num_lines))
print(check_num_lines_of_code(check_num_lines_of_code))
print(check_return(check_return))
# R reformats functions on import (see
# help(get_function_body, package = "cleanr")), so we need 90 characters:
print(check_line_width(check_line_width, maximum = 90))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

