% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_cleanr_options}
\alias{gco}
\alias{get_cleanr_options}
\title{Get options for \pkg{cleanr}}
\usage{
get_cleanr_options(..., remove_names = FALSE, flatten_list = TRUE)

gco(..., remove_names = FALSE, flatten_list = TRUE)
}
\arguments{
\item{...}{See \code{\link{getOption}}.}

\item{remove_names}{[boolean(1)]\cr Remove the names?}

\item{flatten_list}{[boolean(1)]\cr Return a vector?}
}
\value{
a (possibly named) list or a vector.
}
\description{
A convenience function for \code{\link{getOption}}.
}
\note{
\code{gco} is just an alias for \code{get_cleanr_options}.
}
\examples{
get_cleanr_options("max_lines")
get_cleanr_options("max_lines", remove_names = TRUE)
get_cleanr_options("max_lines", flatten_list = TRUE)
get_cleanr_options("max_lines", flatten_list = TRUE, remove_names = TRUE)
get_cleanr_options(flatten_list = TRUE, remove_names = TRUE)
get_cleanr_options(c("max_lines", "max_lines_of_code"))
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}

