% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{get_coldr_options}
\alias{get_coldr_options}
\title{get options for \pkg{coldr}}
\usage{
get_coldr_options(..., remove_names = FALSE, flatten_list = TRUE)
}
\arguments{
\item{...}{see \code{\link{getOption}}/}

\item{remove_names}{[boolean(1)]\cr Remove the names?}

\item{flatten_list}{[boolean(1)]\cr Return a vetcor?}
}
\value{
a (possibly named) list or a vector.
}
\description{
a convenience function for \code{\link{getOption}}.
}
\section{Version}{
 $Id: 01015ff091d53e47fc1caa95805585b6e3911ba5 $
}
\examples{
get_coldr_options('max_lines')
get_coldr_options('max_lines', remove_names = TRUE)
get_coldr_options('max_lines', flatten_list = TRUE)
get_coldr_options('max_lines', flatten_list = TRUE, remove_names = TRUE)
get_coldr_options(flatten_list = TRUE, remove_names = TRUE)
get_coldr_options(c('max_lines', 'max_lines_of_code'))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

