% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_responses.R
\name{get_na_response_rates}
\alias{get_na_response_rates}
\title{Non-responses}
\usage{
get_na_response_rates(data)
}
\arguments{
\item{data}{data set in XML values format}
}
\description{
Non-responses
}
