% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_phonenumbers.R
\name{hash_phonenumbers}
\alias{hash_phonenumbers}
\title{Hash Phone Numbers}
\usage{
hash_phonenumbers(data, PhoneNumber, name, debug = FALSE)
}
\arguments{
\item{name}{}
}
\description{
Hash Phone Numbers
}
