% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_typos.R
\name{identify_typos}
\alias{identify_typos}
\title{Identify Potential Typos in Text Data}
\usage{
identify_typos(data, expected_values = NULL)
}
\arguments{
\item{data}{A character vector containing the treatment values.}

\item{expected_values}{A vector of the two most expected treatment values (optional).}
}
\value{
A data frame of the original values and whether they were flagged as potential typos.
}
\description{
This function flags values that deviate from the two most expected treatment values
or a set of user-defined expected values as potential typos.
}
\examples{
identify_typos(c("d", "w", "wet", "dry", "w", "d", "dr", "w", "d", "d"))
}
