% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_na.R
\name{locate_na}
\alias{locate_na}
\title{Locate NAs}
\usage{
locate_na(data)
}
\arguments{
\item{data}{(tbl_df, df, data.frame) dataframe that the function will use to locate NAs}
}
\value{
a list containing indices of missing values
}
\description{
Locate and return the indices to all missing values within an inputted dataframe.
Each element of the returned list will be a column in a dataframe, which will hold
the row indices of the missing values.
}
