% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_sheet.R
\name{log_sheet}
\alias{log_sheet}
\title{log_sheet}
\usage{
log_sheet(data, question.name, issue, action)
}
\arguments{
\item{data}{this should be survey data in XML value format}

\item{question.name}{this will be the variable name}

\item{issue}{provide further detail on the issue}

\item{action}{type of action required for the issue}
}
\value{
A logbook with flagged observations from \code{\link{log_sheet}}
}
\description{
log_sheet
}
\examples{
log_sheet()
\dontrun{
log_sheet(data = data, question.name = "variable", issue = "issue", action = "check")
}

}
