% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logbook.R
\name{logbook}
\alias{logbook}
\title{logbook}
\usage{
logbook()
}
\arguments{
\item{NA}{No input or argument is required for the function}
}
\value{
An empty logbook from \code{\link{logbook}}
}
\description{
This function will create a blank logbook that will be used to store observations with issues
}
\examples{
logbook()
\dontrun{
cleaning_log <- logbook()
}


}
