% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanR.R
\name{namer}
\alias{namer}
\title{Remove spaces, digits, and any non-letter character from column names}
\usage{
namer(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with modified column names.
}
\description{
Remove spaces, digits, and any non-letter character from column names
}
\examples{
namer(mtcars)
}
