% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_fsl_indicators.R
\name{plot_ridge_distribution}
\alias{plot_ridge_distribution}
\title{plot ridge distribution}
\usage{
plot_ridge_distribution(
  df,
  numeric_cols,
  name_groups = NULL,
  name_units = NULL,
  grouping = NULL,
  file_path = NULL,
  wdth = NULL,
  hght = NULL,
  title_name = NULL
)
}
\arguments{
\item{title_name}{}
}
\description{
plot ridge distribution
}
