% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode}
\alias{recode}
\title{Recode a data set}
\usage{
recode(data, rules)
}
\arguments{
\item{data}{A data set}

\item{rules}{Some rules}
}
\description{
Recode a set of data according to rules. Amazing documentation!
}

