% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_read.R
\name{recode_read}
\alias{recode_read}
\title{Read recoding rules}
\usage{
recode_read(filename, recipes = NULL)
}
\arguments{
\item{filename}{Filename to use}

\item{recipes}{Optional filename for recipes.  If omitted and a
file `recipes.yml` is found, that will be used.}
}
\description{
Read recoding rules from a yaml file.
}

