% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_missing.R
\name{remove_missing}
\alias{remove_missing}
\alias{remove_missing.data.frame}
\title{Remove rows with missing values}
\usage{
remove_missing(x, cols = NULL)

\method{remove_missing}{data.frame}(x, cols = NULL)
}
\arguments{
\item{x}{A data frame.}

\item{cols}{Vector of column names to check for missing values.
If the vector is empty, all columns will be checked.}
}
\value{
A data frame without missing values, since the rows containing them have been removed.
}
\description{
This function is an S3 generic with a method for data frames.
}
\examples{
# As an example, we create a data frame with different columns: the first letter of our names,
# sports, colors and numbers.
# The data frame includes some missing values to demonstrate that this function works correctly.

df <- data.frame(
  names = c("M", "C", "B", "J"),
  sports = c("Football", "Basketball", NA, "Soccer"),
  colors = c("Blue", NA, "White", "Black"),
  numbers = c(1, 2, 3, NA)
)

# Check all columns
remove_missing(df)

# Check only specific columns
remove_missing(df, cols = "sports")
remove_missing(df, cols = c("sports", "numbers"))

}
