% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename.R
\name{rename}
\alias{rename}
\title{Rename columns}
\usage{
rename(data, table, reverse_direction = NULL, allow_missing = FALSE)
}
\arguments{
\item{data}{A data.frame to modify}

\item{table}{A table of \code{from = to} or \code{to = from} mappings}

\item{reverse_direction}{Logical (or NULL) to indicate the
direction of the mapping.  If \code{TRUE}, then \code{table} is
treated as a set of \code{to = from} mappings, while if
\code{FALSE} it is treated as \code{from = to} mappings.  If
\code{NULL} the direction is determined automatically.}

\item{allow_missing}{Allow columns to be missing in \code{data}
(but present \code{table}).  This is not compatible with
\code{reverse_direction = NULL}.}
}
\description{
Bulk rename a number of columns.  Note that this overlaps
significantly with \code{dplyr}'s function of the same name, with
the chief differences being handling of yaml metadata and slightly
odd handling of polarity.  We may just depend on dplyr later.
}

