% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{replace_na}
\alias{replace_na}
\title{Replace NAs}
\usage{
replace_na(data, columns, replace = "mean", remove = FALSE)
}
\arguments{
\item{data}{(tbl_df, df, data.frame) dataframe that the function will use to replace NAs}

\item{columns}{(vector or list) columns that specify which column to replace}

\item{replace}{(character) the method to replace missing values with}

\item{remove}{(logical) set TRUE to remove missing value rows}
}
\value{
A list of tuples where each NAs will be replaced by either
min, max, median or average.
}
\description{
This function replaces na values with either the min, max,
median or average value or removes the rows.
}
