% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{rule.validation}
\alias{rule.validation}
\title{Validate data given a rule set

Throws an error if the data does not confirm to the rule set}
\usage{
rule.validation(data, rules.file)
}
\arguments{
\item{data}{The data to validate}

\item{rule.file}{The yaml file containing the rules}
}

