% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{rule_validation}
\alias{rule_validation}
\title{Validate data given a rule set

Throws an error if the data does not confirm to the rule set}
\usage{
rule_validation(data, rules_file = NULL, rules = NULL)
}
\arguments{
\item{data}{The data to validate}

\item{rules}{List of rules}

\item{rule_file}{Optional: the yaml file containing the rules}
}

