% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{set_coldr_options}
\alias{set_coldr_options}
\title{set options for \pkg{coldr}}
\usage{
set_coldr_options(..., reset = FALSE, overwrite = TRUE)
}
\arguments{
\item{...}{see \code{\link{options}}.}

\item{reset}{[boolean(1)]\cr Reset all \pkg{coldr} options to the package's
defaults?}

\item{overwrite}{[boolean(1)]\cr Overwrite options already set? Is set to
FALSE on package loading to ensure your previously set \pkg{coldr} options
won't get overridden. Just ignore that argument.}
}
\value{
invisible(TRUE)
}
\description{
a convenience function for \code{\link{options}}.
}
\details{
the package loads a couple of options as defaults for its functions.
The defaults are stored in a list element of \code{\link{options}}.
}
\section{Version}{
 $Id: 01015ff091d53e47fc1caa95805585b6e3911ba5 $
}
\examples{
# R.C. Martin's Clean Code recommends monadic argument lists.
set_coldr_options(max_arguments = 1)
# R.C. Martin's Clean Code recommends functions less than 20 lines long.
set_coldr_options(max_lines = 30, max_lines_of_code = 20)
# same as above:
set_coldr_options(list(max_lines = 30, max_lines_of_code = 20))
get_coldr_options(flatten_list = TRUE)
# we delete all options and set some anew
options('coldr' = NULL)
options('coldr' = list(max_lines = 30, max_lines_of_code = 20))
# fill the missing options with the package's defaults:
set_coldr_options(overwrite = FALSE)
get_coldr_options(flatten_list = TRUE)
# reset to the package's defaults:
set_coldr_options(reset = TRUE)
get_coldr_options(flatten_list = TRUE)
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

