% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{tidy_findings}
\alias{tidy_findings}
\title{tidy findings}
\usage{
tidy_findings(findings)
}
\arguments{
\item{findings}{a character vector with possibly some elements reading "TRUE"
or a vector of TRUEs.}
}
\value{
a character vector without any element reading "TRUE" or NULL.
}
\description{
remove TRUE converted to class character from findings.
}
\details{
\code{\link{check_directory}}, \code{\link{check_file}},
\code{\link{check_functions_in_file}} and
\code{\link{check_function_layout}} all collect tryCatch to collect either
TRUE for a check passed or a character holding a conditions message. This
function deletes the TRUEs.
}
\examples{
findings <- c("some signal caught", rep("TRUE", 3))
cleanr:::tidy_findings(findings)
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}
\keyword{internal}

