% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_uuid.R
\name{check_duplicate_uuid}
\alias{check_duplicate_uuid}
\title{Check duplicated uuids}
\usage{
check_duplicate_uuid(data)
}
\arguments{
\item{data}{data in XML values format}
}
\description{
Check duplicated uuids
}
