% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{check_file_layout}
\alias{check_file_layout}
\title{check a file's layout}
\usage{
check_file_layout(path,
  max_file_length = get_cleanr_options("max_file_length"),
  max_file_width = get_cleanr_options("max_file_width"))
}
\arguments{
\item{path}{A path to a file, e.g. "checks.R".}

\item{max_file_length}{The maximum number of lines accepted.}

\item{max_file_width}{The maximum line width accepted.}
}
\value{
invisible(TRUE), but see \emph{Details}.
}
\description{
Check for number of lines and width of lines.
}
\details{
Some reckon a code file should not be too long and that its lines should not
be too wide. On current monitors, 300 lines are about five pages.
A line width of 80 seems a bit \ldots{} outdated, but maybe there's some good
in it.

In case of a fail the function \code{link{throw}}s a
condition of class c("cleanr", "error", "condition").
}
\examples{
print(check_file_layout(system.file("source", "R", "checks.R",
                                    package = "cleanr")))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

