% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{check_file_layout}
\alias{check_file_layout}
\title{Check a File's Layout}
\usage{
check_file_layout(path,
  max_file_length = get_cleanr_options("max_file_length"),
  max_file_width = get_cleanr_options("max_file_width"))
}
\arguments{
\item{path}{Path to the file to be checked.}

\item{max_file_length}{See \code{\link{check_file_length}}.}

\item{max_file_width}{See \code{\link{check_file_width}}.}
}
\value{
invisible(TRUE), but see \emph{Details}.
}
\description{
Run all \code{\link{file_checks}} on a file.
}
\details{
The function catches the messages of "cleanr"-conditions \code{\link{throw}}n
by \code{\link{file_checks}} and, if it caught any, \code{\link{throw}}s
them.
}
\examples{
print(check_file_layout(system.file("source", "R", "checks.R",
                                    package = "cleanr")))
}
\author{
Andreas Dominik Cullmann, <adc-r@arcor.de>
}

