% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{check_function_layout}
\alias{check_function_layout}
\title{Check a Function's Layout}
\usage{
check_function_layout(object,
  max_lines_of_code = get_cleanr_options("max_lines_of_code"),
  max_lines = get_cleanr_options("max_lines"),
  max_arguments = get_cleanr_options("max_arguments"),
  max_nesting_depth = get_cleanr_options("max_nesting_depth"),
  max_line_width = get_cleanr_options("max_line_width"),
  check_return = get_cleanr_options("check_return"))
}
\arguments{
\item{object}{The function to be checked.}

\item{max_lines_of_code}{See \code{\link{check_num_lines_of_code}}.}

\item{max_lines}{See \code{\link{check_num_lines}}.}

\item{max_arguments}{See \code{\link{check_num_arguments}}.}

\item{max_nesting_depth}{See \code{\link{check_nesting_depth}}.}

\item{max_line_width}{See \code{\link{check_line_width}}.}

\item{check_return}{See \code{\link{check_return}}.}
}
\value{
invisible(TRUE), but see \emph{Details}.
}
\description{
Run all \code{\link{function_checks}} on a function.
}
\details{
The functions catches the messages of "cleanr"-conditions
\code{\link{throw}}n by \code{\link{function_checks}} and, if it caught any,
\code{\link{throw}}s them.
}
\examples{
print(check_function_layout(check_num_lines))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}

