% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_variables.R
\name{compare_variables}
\alias{compare_variables}
\title{compare_variables}
\usage{
compare_variables(template, newdat, threshold = 1.5)
}
\arguments{
\item{newdat}{}
}
\value{

}
\description{
compare_variables
}
