% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_variables.R
\name{compare_variables}
\alias{compare_variables}
\title{compare_variables}
\usage{
compare_variables(template = check, newdat = dirty_data,
  threshold = 1.5, k = 2)
}
\arguments{
\item{threshold}{: theshold to detect outlyers. Not implemented}

\item{k}{: Fuzzy matching precision for categories}
}
\description{
compare_variables
}
