% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers in numeric data}
\usage{
detect_outliers(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A logical vector indicating outliers.
}
\description{
Detects outliers using a simple statistical rule.
}
