% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_not_false}
\alias{is_not_false}
\title{Test if an Object is not False}
\usage{
is_not_false(x, null_is_false = TRUE, ...)
}
\arguments{
\item{x}{The object to be tested.}

\item{null_is_false}{[boolean(1)]\cr Should NULL be treated as FALSE?}

\item{...}{Parameters passed to \code{\link{exists}}. See Details and
Examples.}
}
\value{
TRUE if the object is set to something different than FALSE, FALSE
otherwise.
}
\description{
See \code{\link[fritools:is_not_false]{fritools::is_not_false}}.
}
\keyword{internal}
