% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_internal_functions}
\alias{load_internal_functions}
\title{Load a Package's Internals}
\usage{
load_internal_functions(package, ...)
}
\arguments{
\item{package}{[character(1)]\cr The name of the package as a string.}

\item{...}{Arguments passed to \code{\link{ls}}, all.names = TRUE could be a
good idea.}
}
\value{
invisible(TRUE)
}
\description{
See
\code{
\link[fritools:load_internal_functions]{fritools::load_internal_functions}
}.
}
\seealso{
\code{\link[codetools:checkUsageEnv]{checkUsageEnv in codetools}}.
}
\keyword{internal}
