% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na.R
\name{replace_na}
\alias{replace_na}
\title{Replace NAs}
\usage{
replace_na(data)
}
\arguments{
\item{data}{(tbl_df, df, data.frame) dataframe that the function will use to locate NAs}
}
\value{
A list of tuples where each NAs will be replaced by either
min, max, median or average.
}
\description{
This function replaces na values with either the min, max,
median or average value or removes the rows.
}
