% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_na.R
\name{replace_na}
\alias{replace_na}
\title{Locate NAs}
\usage{
replace_na(data)
}
\arguments{
\item{data}{(input_df, df, data.frame) input dataframe that the replace_na function will use to replace NAs with with either the min, max, #'median or average value or removes the rows}
}
\value{
A list of tuples where each NAs will be replaced by either #'min, max, median or average.
}
\description{
The function will first find the columns that contains the NAs and return the indices of all the NAs present in the dataframe that we used as an input.
Each tuple in the list will represent  the indices of a NA in the dataframe.
}
