% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_time.R
\name{survey_time}
\alias{survey_time}
\title{survey time check}
\usage{
survey_time(df, time_min, time_max)
}
\arguments{
\item{df}{data frame in XML Values format}

\item{time_min}{minimum time to complete the survey}

\item{time_max}{maximum time to complete the survey}
}
\description{
This function will take dataset with start and end times and compute the survey completion time in minutes.
}
