% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{throw}
\alias{throw}
\title{throw a condition}
\usage{
throw(message_string, system_call = sys.call(-1), ...)
}
\arguments{
\item{message_string}{The message to be thrown.}

\item{system_call}{The call to be thrown.}

\item{...}{Arguments to be passed to \code{\link{structure}}.}
}
\value{
FALSE. But it doesn't return anything, it stops with a
condition of class c('coldr', 'error', 'condition').
}
\description{
throws a condition of class c('coldr', 'error', 'condition').
}
\details{
We use this condition as an error dedictated to \pkg{coldr}.
}
\section{Version}{
 $Id: 01015ff091d53e47fc1caa95805585b6e3911ba5 $
}
\examples{
tryCatch(coldr:::throw('Hello error!'), coldr = function(e) return(e))
}
\author{
Dominik Cullmann, <dominik.cullmann@forst.bwl.de>
}
\keyword{internal}

